refactor(filemanager): comments, move event ingestion to ARCHITECTURE.md, merge collector_builder.rs with collector.rs
