+++ lib/workload/stateful/filemanager/database/queries/ingester/aws/update_reordered_for_created.sql
@@ -106,7 +106,7 @@ select
     created_date as event_time,
     last_modified_date,
     e_tag,
-    storage_class as "storage_class: Option<StorageClass>",
+    storage_class as "storage_class?: StorageClass",
     version_id,
     created_sequencer as sequencer,
     number_reordered,
