+++ lib/workload/stateful/filemanager/filemanager/README.md
@@ -22,38 +22,6 @@ contain thin wrappers around clients, and any logic should be elsewhere to ensur
 
 Generally code that belongs to a particular cloud service should be put in its own module (e.g. AWS code goes in an `aws` module).
 
-## Event ingestion
-
-The filemanager determines the location of objects in cloud storage by ingesting events like the [AWS S3 events][s3-events].
-However, events for most cloud services only guarantee that events are received at least once. Duplicate events may be received
-or events could be received out of order, and the filemanager should handle this. This needs to occur both in application
-code and at the database level because events can be out of order across Lambda function calls. The aim of the filemanager
-is to maintain a database state that is as correct as possible at the time an event is received.
-
-To detect out of order and duplicate events, the filemanager relies on the fields within the events received from AWS.
-These fields are the `bucket`, `key`, `version_id`, and `sequencer` values. For an S3 object, when the `bucket`, `key`
-and `version_id` are the same, the sequencer value determines the ordering of events. An event occurs earlier if it's
-sequencer value is smaller than another event's sequencer value. Duplicate events have the same `bucket`, `key`,
-`version_id` and `sequencer` values for an event type.
-
-### Duplicate events
-
-Within the application code, duplicate events are removed in the [events] module by matching on the `bucket`, `key`,
-`version_id` and `sequencer` values.
-
-At the database level, duplicate events are removed using a unique constraint on the same values and an
-`on conflict on constraint` statement.
-
-### Out of order events
-
-Within the application code, out of order events are removed within the [events] module by comparing sequencer values.
-
-At the database level, events are processed as they arrive. For each object in the database, the sequencer value is
-recorded. When an event is inserted, it is first checked to see if it belongs to an already existing object, i.e. whether
-there are any objects with sequencer values that are greater (for created events) or lower (for deleted events) than the
-existing sequencer. If this condition is met, then the existing object is updated, and the old event is returned by the
-database to be reinserted. If it is not met, then the event is inserted normally.
-
 [clients]: src/clients
 [database]: src/database
 [events]: src/events
